+++ package.json
@@ -146,7 +146,7 @@
     "@types/js-yaml": "^3.12.5",
     "@types/mocha": "^7.0.2",
     "@types/node": "^13.13.15",
-    "@types/vscode": "^1.47.0",
+    "@types/vscode": "^1.52.0",
     "@typescript-eslint/eslint-plugin": "^2.30.0",
     "@typescript-eslint/parser": "^2.30.0",
     "eslint": "^6.8.0",
